+++ Pandora/Paradigm/Primary/Functor/Conclusion.hs
@@ -9,7 +9,7 @@ import Pandora.Pattern.Functor.Applicative (Applicative ((<*>)), Applicative_ (m
 import Pandora.Pattern.Functor.Traversable (Traversable ((->>)), Traversable_ ((-<<--)))
 import Pandora.Pattern.Functor.Bindable (Bindable ((>>=)), Bindable_ (join_))
 import Pandora.Pattern.Functor.Monad (Monad)
-import Pandora.Pattern.Functor.Bivariant (Bivariant ((<->)), Bivariant_ ((-<->-)))
+import Pandora.Pattern.Functor.Bivariant (Bivariant_ ((-<->-)))
 import Pandora.Pattern.Object.Setoid (Setoid ((==)))
 import Pandora.Pattern.Object.Chain (Chain ((<=>)))
 import Pandora.Pattern.Object.Semigroup (Semigroup ((+)))
@@ -76,9 +76,6 @@ instance Bindable_ (Conclusion e) (->) where
 	join_ (Failure y) = Failure y
 
 instance Monad (Conclusion e) where
-
-instance Bivariant Conclusion where
-	f <-> g = conclusion # Failure . f # Success . g
 
 instance Bivariant_ Conclusion (->) (->) (->) where
 	f -<->- g = conclusion # Failure . f # Success . g