+++ Pandora/Paradigm/Primary/Functor/Constant.hs
@@ -1,12 +1,12 @@
 module Pandora.Paradigm.Primary.Functor.Constant where
 
-import Pandora.Pattern.Category (($))
+import Pandora.Pattern.Category ((.), ($))
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$>)), Covariant_ ((-<$>-)))
 import Pandora.Pattern.Functor.Contravariant (Contravariant ((>$<)))
 import Pandora.Pattern.Functor.Invariant (Invariant ((<$<)))
 import Pandora.Pattern.Functor.Pointable (Pointable (point))
 import Pandora.Pattern.Functor.Traversable (Traversable ((->>)))
-import Pandora.Pattern.Functor.Bivariant (Bivariant ((<->)))
+import Pandora.Pattern.Functor.Bivariant (Bivariant_ ((-<->-)))
 import Pandora.Pattern.Object.Setoid (Setoid ((==)))
 import Pandora.Pattern.Object.Chain (Chain ((<=>)))
 import Pandora.Pattern.Object.Semigroup (Semigroup ((+)))
@@ -17,6 +17,7 @@ import Pandora.Pattern.Object.Semilattice (Infimum ((/\)), Supremum ((\/)))
 import Pandora.Pattern.Object.Lattice (Lattice)
 import Pandora.Pattern.Object.Group (Group (invert))
 import Pandora.Paradigm.Primary.Functor.Function ()
+import Pandora.Paradigm.Primary.Transformer.Flip (Flip (Flip))
 
 newtype Constant a b = Constant a
 
@@ -25,6 +26,9 @@ instance Covariant (Constant a) where
 
 instance Covariant_ (Constant a) (->) (->) where
 	_ -<$>- Constant x = Constant x
+
+instance Covariant_ (Flip Constant b) (->) (->) where
+	f -<$>- Flip (Constant x) = Flip . Constant $ f x
 
 instance Contravariant (Constant a) where
 	_ >$< Constant x = Constant x
@@ -35,8 +39,8 @@ instance Invariant (Constant a) where
 instance Traversable (Constant a) where
 	Constant x ->> _ = point $ Constant x
 
-instance Bivariant Constant where
-	f <-> _ = \(Constant x) -> Constant $ f x
+instance Bivariant_ Constant (->) (->) (->)where
+	f -<->- _ = \(Constant x) -> Constant $ f x
 
 instance Setoid a => Setoid (Constant a b) where
 	Constant x == Constant y = x == y