+++ Pandora/Pattern/Functor/Divisible.hs
@@ -1,7 +1,7 @@
 module Pandora.Pattern.Functor.Divisible where
 
 import Pandora.Pattern.Functor.Contravariant (Contravariant)
-import Pandora.Pattern.Functor.Bivariant (Bivariant)
+import Pandora.Pattern.Functor.Bivariant (Bivariant_)
 import Pandora.Paradigm.Primary.Functor.Product (type (:*:))
 
 infixr 5 >*<
@@ -10,5 +10,5 @@ class Contravariant t => Divisible t where
 	{-# MINIMAL (>*<) #-}
 	(>*<) :: t b -> t c -> t (b :*: c)
 
-class Bivariant v => Divisible_ t v source target where
-	divide :: Bivariant v => source r (v a b) -> target (v (t a) (t b)) (t r)
+class Bivariant_ v source source target => Divisible_ t v source target where
+	divide :: source r (v a b) -> target (v (t a) (t b)) (t r)